Devq

+++ src/Engine/EngineApp/EngineApp.cpp
@@ -34,7 +34,7 @@ void EngineApp::run() {
         lastFrameTime = runTime;
         nFrames++;
         if (runTime - lastFpsTime >= 1.0) {
-            fps = (double)nFrames;
+            fps = nFrames;
             nFrames = 0;
             lastFpsTime = runTime;
         }
